Terrier compiles again after fixing missing includes downstream.
